requirements.txt
flask
tensorflow
opencv-python
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib